+++ packages/cosmwasm-stargate/src/queries/wasm.spec.ts
@@ -84,7 +84,7 @@ async function instantiateContract(
           beneficiary: beneficiaryAddress,
         }),
       ),
-      initFunds: transferAmount ? [...transferAmount] : [],
+      funds: transferAmount ? [...transferAmount] : [],
     }),
   };
   const fee: StdFee = {
@@ -109,7 +109,7 @@ async function executeContract(
       sender: alice.address0,
       contract: contractAddress,
       msg: toAscii(JSON.stringify(msg)),
-      sentFunds: [],
+      funds: [],
     }),
   };
   const fee: StdFee = {
@@ -219,6 +219,7 @@ describe("WasmExtension", () => {
         creator: alice.address0,
         label: "my escrow",
         admin: "",
+        ibcPortId: "",
       });
 
       const { contractInfo } = await client.unverified.wasm.getContractInfo(myAddress);
@@ -228,6 +229,7 @@ describe("WasmExtension", () => {
         creator: alice.address0,
         label: "my escrow",
         admin: "",
+        ibcPortId: "",
       });
       expect(contractInfo.admin).toEqual("");
     });
@@ -419,9 +421,11 @@ describe("WasmExtension", () => {
         assertDefined(result.data);
         const msgData = fromOneElementArray(result.data);
         expect(msgData.msgType).toEqual("instantiate");
-        expect(MsgInstantiateContractResponse.decode(msgData.data)).toEqual(
-          MsgInstantiateContractResponse.fromPartial({ address: contractAddress }),
-        );
+        const response = MsgInstantiateContractResponse.decode(msgData.data);
+        expect({
+          ...response,
+          data: new Uint8Array(), // workaround for https://github.com/stephenh/ts-proto/issues/237
+        }).toEqual(MsgInstantiateContractResponse.fromPartial({ address: contractAddress }));
 
         const balanceUcosm = await client.bank.balance(contractAddress, "ucosm");
         expect(balanceUcosm).toEqual(transferAmount[0]);