+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # CHANGELOG
 
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to
+[Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+### Changed
+
+- @cosmjs/cosmwasm-stargate: Codec adapted to support wasmd 0.16. Older versions
+  of wasmd are not supported anymore.
+
 ## [0.24.0] - 2021-03-11
 
 - @cosmjs/cosmwasm: This package is now deprecated. The same functionality is
@@ -242,4 +253,5 @@
   `FeeTable`. @cosmjs/cosmwasm has its own `FeeTable` with those properties.
 - @cosmjs/sdk38: Rename package to @cosmjs/launchpad.
 
+[unreleased]: https://github.com/cosmos/cosmjs/compare/v0.24.0...HEAD
 [0.24.0]: https://github.com/cosmos/cosmjs/compare/v0.23.0...v0.24.0